+++ paddleformers/trainer/trainer_callback.py
@@ -25,9 +25,9 @@ from dataclasses import dataclass
 from typing import Dict, List, Optional, Union
 
 import numpy as np
-from moe_utils import offload, reload
 from tqdm.auto import tqdm
 
+from ..transformers.moe_utils import offload, reload
 from ..utils.log import logger
 from .trainer_utils import IntervalStrategy, has_length
 from .training_args import TrainingArguments
